Added the functions as requested
Can get rid of the print statements inside if needed. Also, sqlite handler is imported based on importing from the root Scripts directory. This does not allow the module to be run as a standalone. To get around this you will want a copy of SQLitehandler.py from Scripts into the Util folder for the import to work.